Refactor: move `load_log_ids()` from `RaftState` to `LogIdList`
This method does not need to be a method of `RaftState`.
